requirements.txt
altair==5.5.0
attrs==20.3.0
aws-cfn-bootstrap==2.0
awscli==2.23.11
awscrt==0.23.8
Babel==2.9.1
biopython==1.85
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.14.5
chardet==4.0.0
charset-normalizer==3.4.1
chevron==0.13.1
click==8.1.8
cloud-init==22.2.2
colorama==0.4.4
configobj==5.0.6
contourpy==1.3.0
cryptography==36.0.1
cycler==0.12.1
dbus-python==1.2.18
distro==1.5.0
docutils==0.16
ec2-hibinit-agent==1.0.8
et_xmlfile==2.0.0
fonttools==4.57.0
gitdb==4.0.12
GitPython==3.1.44
gpg==1.15.1
idna==2.10
importlib_resources==6.5.2
japanize-matplotlib==1.1.3
Jinja2==2.11.3
jmespath==0.10.0
jsonpatch==1.21
jsonpointer==2.0
jsonschema==3.2.0
kiwisolver==1.4.7
libcomps==0.1.20
lockfile==0.12.2
lxml==5.3.2
MarkupSafe==1.1.1
matplotlib==3.9.4
narwhals==1.34.0
netifaces==0.10.6
numpy==2.0.2
oauthlib==3.0.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
ply==3.11
prettytable==0.7.2
prompt-toolkit==3.0.24
protobuf==5.29.4
pyarrow==19.0.1
pyCirclize==1.9.0
pycparser==2.20
pydeck==0.9.1
pyparsing==3.2.3
pyrsistent==0.17.3
pyserial==3.4
PySocks==1.7.1
python-daemon==2.3.0
python-dateutil==2.9.0.post0
python-pptx==1.0.2
pytz==2022.7.1
PyYAML==5.4.1
release-notification==1.2
requests==2.32.3
rpm==4.16.1.3
ruamel.yaml==0.16.6
ruamel.yaml.clib==0.1.2
selinux==3.4
sepolicy==3.4
setools==4.4.1
six==1.15.0
smmap==5.0.2
streamlit==1.44.1
support-info==1.0
systemd-python==235
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing_extensions==4.13.1
tzdata==2025.2
urllib3==1.25.10
watchdog==6.0.0
wcwidth==0.2.5
XlsxWriter==3.2.2
zipp==3.21.0
